extract legacy event bus action to dedicated file
for easier refactoring in future
